+++ net/dll/seal/contextdata.h
@@ -13,7 +13,7 @@
 
 SEALDLL HRESULT SEALCALL ContextData_Destroy(void* thisptr);
 
-SEALDLL HRESULT SEALCALL ContextData_TotalCoeffModulus(void* thisptr, int* count, uint64_t* total_coeff_modulus);
+SEALDLL HRESULT SEALCALL ContextData_TotalCoeffModulus(void* thisptr, uint64_t* count, uint64_t* total_coeff_modulus);
 
 SEALDLL HRESULT SEALCALL ContextData_TotalCoeffModulusBitCount(void* thisptr, int* bit_count);
 
@@ -21,15 +21,15 @@ SEALDLL HRESULT SEALCALL ContextData_Parms(void* thisptr, void** parms);
 
 SEALDLL HRESULT SEALCALL ContextData_Qualifiers(void* thisptr, void** epq);
 
-SEALDLL HRESULT SEALCALL ContextData_CoeffDivPlainModulus(void* thisptr, int* count, uint64_t* coeff_div);
+SEALDLL HRESULT SEALCALL ContextData_CoeffDivPlainModulus(void* thisptr, uint64_t* count, uint64_t* coeff_div);
 
 SEALDLL HRESULT SEALCALL ContextData_PlainUpperHalfThreshold(void* thisptr, uint64_t* puht);
 
-SEALDLL HRESULT SEALCALL ContextData_PlainUpperHalfIncrement(void* thisptr, int* count, uint64_t* puhi);
+SEALDLL HRESULT SEALCALL ContextData_PlainUpperHalfIncrement(void* thisptr, uint64_t* count, uint64_t* puhi);
 
-SEALDLL HRESULT SEALCALL ContextData_UpperHalfThreshold(void* thisptr, int* count, uint64_t* uht);
+SEALDLL HRESULT SEALCALL ContextData_UpperHalfThreshold(void* thisptr, uint64_t* count, uint64_t* uht);
 
-SEALDLL HRESULT SEALCALL ContextData_UpperHalfIncrement(void* thisptr, int* count, uint64_t* uhi);
+SEALDLL HRESULT SEALCALL ContextData_UpperHalfIncrement(void* thisptr, uint64_t* count, uint64_t* uhi);
 
 SEALDLL HRESULT SEALCALL ContextData_NextContextData(void* thisptr, void** next_data);
